Adding base_url parameter to Github Actions config
The main usage of this actions is serving Jekyll sites from GH Pages. As in majority of cases this sites get served from sub domain (by repo name) it would useful to be able to set base_url for this sites. Sure it's probably can be done with different configs (_config_production.yml etc) but having this option makes it easier.